Smartphone Based Seizure Detection: Android App
This android application was made as a part of Philly Codefest 2015
Competition.
The app aims to acquire real time data from the accelerometer and
gyroscopres of the smartphone and feed it to the Microsoft Azure
platform for further processing.
Wherein, the Azure's Machine Learning framework has been used to
classify an activity as a Seizure or Not a Seizure.
If it is a seizure then the closed ones of the app user as well as the
emergency services will be notified through the app.

Any suggestions/comments regarding this are welcome.
Contact me on : sagarpatel1911@gmail.com